% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_bspbss.R
\name{init_bspbss}
\alias{init_bspbss}
\title{Title}
\usage{
init_bspbss(
  X,
  xgrid,
  mask = rep(1, nrow(xgrid)),
  q = 2,
  dens = 0.5,
  ker_par = c(0.05, 20),
  num_eigen = 500,
  noise = 0
)
}
\arguments{
\item{noise}{}
}
\value{

}
\description{
Title
}
