% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_bspbss.R
\name{mcmc_bspbss_pickup}
\alias{mcmc_bspbss_pickup}
\title{Title}
\usage{
mcmc_bspbss_pickup(
  X,
  res,
  prior,
  kernel,
  ep = 0.5,
  lr = 0.1,
  subsample_n,
  subsample_p,
  n.iter,
  n.burn_in,
  thin = 1,
  show_step,
  noise = 0.1
)
}
\arguments{
\item{show_step}{}
}
\value{

}
\description{
Title
}
