% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nii.R
\name{output_nii}
\alias{output_nii}
\title{Write a NIfTI file.}
\usage{
output_nii(X, nii, xgrid, file = NULL, std = TRUE, thres = 0)
}
\arguments{
\item{X}{Data matrix with n rows (sample) and p colums (pixel).}

\item{nii}{a reference NIfTI-class object, representing a image with p0 voxels.}

\item{xgrid}{Cordinate matrix with p0 rows (voxel) and d columns (dimension). p0 >= p.}

\item{file}{The name of the file to be saved.}

\item{std}{If TRUE, standarize each row of X.}

\item{thres}{Quantile to threshold each row of X.}
}
\description{
This function saves a data matrix into a NIfTI file.
}
