% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_nii.R
\name{pre_nii}
\alias{pre_nii}
\title{Read a NIfTI file}
\usage{
pre_nii(nii, mask)
}
\arguments{
\item{nii_file_list}{A list of NIfTI file(s).}

\item{mask_file}{Path to the mask file.}
}
\value{

}
\description{
This function transforms NIfTI file(s) into a data matrix.
}
