% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_bspbss.R
\name{init_bspbss}
\alias{init_bspbss}
\title{Initial values}
\usage{
init_bspbss(
  X,
  xgrid,
  mask = rep(1, nrow(xgrid)),
  q = 2,
  dens = 0.5,
  kernel = "gaussian",
  ker_par = c(0.05, 20),
  num_eigen = 500,
  noise = 0
)
}
\arguments{
\item{X}{Data matrix with n rows (sample) and p columns (voxel).}

\item{xgrid}{Cordinate matrix with p0 rows (voxel) and d columns (dimension). p0 >= p.}

\item{mask}{Logical vector of length p0, with p elements being equal to 1 and p0-p elements being equal to 0.}

\item{q}{Number of latent sources.}

\item{dens}{The initial density level (between 0 and 1) of the latent sources.}

\item{kernel}{"gaussian" or "matern".}

\item{ker_par}{2-dimensional vector (a,b) with a>0, b>0, specifing the parameters in the modified exponetial squared kernel.}

\item{num_eigen}{Number of the eigen functions.}

\item{noise}{Gaussian noise added to the initial latent sources, with mean 0 and standard deviation being noise * sd(S0),
where sd(S0) is the standard deviation of the initial latent sources.}
}
\value{
List includes the initial values, priors and eigen functions/eigen values of the kernel.
}
\description{
Generate initial values, set up priors and perform kernel decomposition
for runing MCMC algorithm.
}
\examples{

}
