% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplot2D.R
\name{levelplot2D}
\alias{levelplot2D}
\title{levelplot for 2D images.}
\usage{
levelplot2D(
  S,
  lim = c(min(S), max(S)),
  xgrid,
  mask = rep(1, ncol(S)),
  color = bluered(100),
  path,
  name
)
}
\arguments{
\item{S}{Data matrix with q rows (sample) and p colums (pixel).}

\item{lim}{2-dimensional vector specifying the limits for the data.}

\item{xgrid}{Cordinate matrix with p0 rows (voxel) and d columns (dimension).p0 >= p.}

\item{mask}{Logical vector of length p0, with p elements being equal to 1 and p0-p elements being equal to 0.}

\item{color}{Colorbar.}

\item{path}{Path that images are saved at, ended with a "/" or "\".}

\item{name}{File names.}
}
\value{

}
\description{
The function plots 2D images for a data matrix.
}
