% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_bspbss.R
\name{mcmc_bspbss}
\alias{mcmc_bspbss}
\title{MCMC algorithm for Bayesian spatial blind source separation
with the thresholded Gaussian Process prior.}
\usage{
mcmc_bspbss(
  X,
  init,
  prior,
  kernel,
  n.iter,
  n.burn_in,
  thin = 1,
  show_step,
  ep = 0.1,
  lr = 0.01,
  decay = 0.01,
  subsample_n = 0.5,
  subsample_p = 0.5
)
}
\arguments{
\item{X}{Data matrix with n rows (sample) and p columns (voxel).}

\item{init}{List of initial values, see \code{init_mcmc}.}

\item{prior}{List of priors, see \code{init_mcmc}.}

\item{kernel}{List including eigenvalues and eigenfunctions of the kernel, see \code{init_TGP}.}

\item{n.iter}{Total iterations in MCMC.}

\item{n.burn_in}{Number of burn-in.}

\item{thin}{Thining interval.}

\item{show_step}{Frequency for printing the current number of iterations.}

\item{ep}{Approximation parameter.}

\item{lr}{Per-batch learning rate.}

\item{subsample_n}{Size of a mini-batch for samples.}

\item{subsample_p}{Size of a mini-batch for voxels.}
}
\value{
List that contains MCMC samples for: A, b, sigma, zeta and tau.
}
\description{
Performan MCMC algorithm to draw samples from a Bayesian spatial blind source separation
model.
}
