% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_nii.R
\name{output_nii}
\alias{output_nii}
\title{Write a NIfTI file.}
\usage{
output_nii(X, nii, xgrid, filename = NULL, std = TRUE, thres = 0)
}
\arguments{
\item{X}{Data matrix with n rows (sample) and p colums (pixel).}

\item{nii}{a reference NIfTI-class object, representing a image with p0 voxels.}

\item{xgrid}{Cordinate matrix with p0 rows (voxel) and d columns (dimension). p0 >= p.}

\item{mask}{Logical vector of length p0, with p elements being equal to 1 and p0-p elements being equal to 0.}

\item{file}{The name to the file to be saved.}
}
\value{

}
\description{
This function saves a data matrix into a NIfTI file.
}
