% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_nii.R
\name{pre_nii}
\alias{pre_nii}
\title{Read a NIfTI file}
\usage{
pre_nii(nii, mask)
}
\arguments{
\item{nii}{Path to the NIfTI file.}

\item{mask}{Path to the mask file.}
}
\description{
This function transforms NIfTI file into a data matrix.
}
